Move service processing into a concern
